requirement.txt
Flask
transformers
pandas
requests
openai
python-dotenv
Mastodon.py
atproto
regex